requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==4.2.16
django-request==1.6.3
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
